5cat_file/requirements.txt
google-cloud-storage
google-auth
google-cloud-secret-manager
anthropic
Pillow 